test: replace Popconfirm part test with testing lib (#37703)
* test: replace Popconfirm part test with testing lib

* test: Update test case
